Move `HsBindgen.Hs` to `HsBindgen.Backend.Hs`
