+++ hs-bindgen/src-internal/HsBindgen/Pipeline/Lib.hs
@@ -26,14 +26,14 @@ import HsBindgen.Backend.Artefact.PP.Translation qualified as Backend.PP
 import HsBindgen.Backend.Artefact.Test qualified as GenTests
 import HsBindgen.Backend.Artefact.TH.Translation qualified as Backend.TH
 import HsBindgen.Backend.Extensions
+import HsBindgen.Backend.Hs.AST qualified as Hs
+import HsBindgen.Backend.Hs.Translation qualified as Hs
 import HsBindgen.BindingSpec
 import HsBindgen.Config (Config (..))
 import HsBindgen.Frontend
 import HsBindgen.Frontend.AST.External qualified as C
 import HsBindgen.Frontend.RootHeader
 import HsBindgen.Guasi
-import HsBindgen.Hs.AST qualified as Hs
-import HsBindgen.Hs.Translation qualified as Hs
 import HsBindgen.Imports
 import HsBindgen.ModuleUnique
 import HsBindgen.SHs.AST qualified as SHs
